Classify supported views
