+++ packages/notifi-react-example/src/AcalaWalletContextProvider.tsx
@@ -254,7 +254,6 @@ export const AcalaWalletContextProvider: FC<PropsWithChildren<{}>> = ({
 
   return (
     <AcalaWalletContext.Provider value={values}>
-      <AcalaConnectButton />
       {children}
     </AcalaWalletContext.Provider>
   );
