+++ packages/notifi-react-example/src/NotifiCard/PolkadotCard.tsx
@@ -1,0 +1,10 @@
+import { AcalaConnectButton } from '../AcalaWalletContextProvider';
+
+export const PolkadotCard = () => {
+  return (
+    <div className="container">
+      <h1>Notifi Card: Polkadot</h1>
+      <AcalaConnectButton />
+    </div>
+  );
+};
