+++ packages/notifi-react-example/src/SolanaWalletProvider.tsx
@@ -46,11 +46,7 @@ export const SolanaWalletProvider: FC<PropsWithChildren<{}>> = ({
   return (
     <ConnectionProvider endpoint={endpoint}>
       <WalletProvider wallets={wallets} autoConnect>
-        <WalletModalProvider>
-          <WalletMultiButton />
-          <WalletDisconnectButton />
-          {children}
-        </WalletModalProvider>
+        <WalletModalProvider>{children}</WalletModalProvider>
       </WalletProvider>
     </ConnectionProvider>
   );
